requirements.txt
Flask==3.0.3
qrcode==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
Werkzeug==3.0.3
itsdangerous==2.2.0
pytest==8.2.0
gunicorn==22.0.0
Flask-Login==0.6.3
Flask-Limiter==3.5.1
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0